feat: 로그인/비로그인별 서비스 구분

+++ src/main/java/custom/capstone/domain/inquiry/application/AnswerService.java
@@ -8,11 +8,13 @@ import custom.capstone.domain.inquiry.dto.request.AnswerSaveRequestDto;
 import custom.capstone.domain.inquiry.dto.request.AnswerUpdateRequestDto;
 import custom.capstone.domain.inquiry.dto.response.AnswerSaveResponseDto;
 import custom.capstone.domain.inquiry.dto.response.AnswerUpdateResponseDto;
+import custom.capstone.domain.inquiry.exception.AnswerInvalidException;
 import custom.capstone.domain.inquiry.exception.AnswerNotFoundException;
 import custom.capstone.domain.inquiry.exception.InquiryNotFoundException;
+import custom.capstone.domain.members.dao.MemberRepository;
 import custom.capstone.domain.members.domain.Member;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,21 +23,23 @@ import org.springframework.transaction.annotation.Transactional;
 public class AnswerService {
     private final InquiryRepository inquiryRepository;
     private final AnswerRepository answerRepository;
+    private final MemberRepository memberRepository;
 
     /**
      * 답변 등록
      */
     @Transactional
     public AnswerSaveResponseDto saveAnswer(
-            final Member member,
-            final Long inquiryId,
+            final String loginEmail,
             final AnswerSaveRequestDto requestDto
     ) {
-        final Inquiry inquiry = inquiryRepository.findById(inquiryId)
+        final Member admin = getValidAdmin(loginEmail);
+
+        final Inquiry inquiry = inquiryRepository.findById(requestDto.inquiryId())
                 .orElseThrow(InquiryNotFoundException::new);
 
         final Answer answer = Answer.builder()
-                .admin(member.getEmail())
+                .admin(admin.getNickname())
                 .content(requestDto.content())
                 .build();
 
@@ -50,14 +54,16 @@ public class AnswerService {
      */
     @Transactional
     public AnswerUpdateResponseDto updateAnswer(
-            final Member member,
+            final String loginEmail,
             final Long answerId,
             final AnswerUpdateRequestDto requestDto
     ) {
+        final Member admin = getValidAdmin(loginEmail);
+
         final Answer answer = answerRepository.findById(answerId)
                 .orElseThrow(AnswerNotFoundException::new);
 
-        checkEqualMember(member, answer);
+        checkEqualAdmin(admin, answer);
 
         answer.update(requestDto.content());
 
@@ -67,18 +73,26 @@ public class AnswerService {
     /**
      * 답변 삭제
      */
-    public void deleteAnswer(final Member member, final Long answerId) {
+    public void deleteAnswer(final String loginEmail, final Long answerId) {
+        final Member admin = getValidAdmin(loginEmail);
+
         final Answer answer = answerRepository.findById(answerId)
                 .orElseThrow(AnswerNotFoundException::new);
 
-        checkEqualMember(member, answer);
+        checkEqualAdmin(admin, answer);
 
         answerRepository.delete(answer);
     }
 
-    // 관리자가 맞는지 확인
-    private void checkEqualMember(final Member member, final Answer answer) {
-        if (!answer.getAdmin().equals(member.getEmail()))
-            throw new AccessDeniedException("해당 권한이 없습니다.");
+    // 관리자인지 확인
+    private Member getValidAdmin(final String loginEmail) {
+        return memberRepository.findByEmail(loginEmail)
+                .orElseThrow(() -> new UsernameNotFoundException("관리자를 찾을 수 없습니다."));
+    }
+
+    // 답변한 관리자가 맞는지 확인
+    private void checkEqualAdmin(final Member admin, final Answer answer) {
+        if (!answer.getAdmin().equals(admin))
+            throw new AnswerInvalidException();
     }
 }
