+++ src/main/java/custom/capstone/domain/notice/application/NoticeService.java
@@ -1,6 +1,6 @@
 package custom.capstone.domain.notice.application;
 
-import custom.capstone.domain.members.application.MemberService;
+import custom.capstone.domain.members.dao.MemberRepository;
 import custom.capstone.domain.members.domain.Member;
 import custom.capstone.domain.notice.dao.NoticeRepository;
 import custom.capstone.domain.notice.domain.Notice;
@@ -10,10 +10,12 @@ import custom.capstone.domain.notice.dto.response.NoticeListResponseDto;
 import custom.capstone.domain.notice.dto.response.NoticeResponseDto;
 import custom.capstone.domain.notice.dto.response.NoticeSaveResponseDto;
 import custom.capstone.domain.notice.dto.response.NoticeUpdateResponseDto;
+import custom.capstone.domain.notice.exception.NoticeInvalidException;
 import custom.capstone.domain.notice.exception.NoticeNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,19 +23,19 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class NoticeService {
     private final NoticeRepository noticeRepository;
-    private final MemberService memberService;
+    private final MemberRepository memberRepository;
 
     /**
      * 공지사항 등록
      */
     @Transactional
-    public NoticeSaveResponseDto saveNotice(final NoticeSaveRequestDto requestDto) {
-        final Member member = memberService.findById(requestDto.memberId());
+    public NoticeSaveResponseDto saveNotice(final String loginEmail, final NoticeSaveRequestDto requestDto) {
+        final Member admin = getValidAdmin(loginEmail);
 
         final Notice notice = Notice.builder()
                 .title(requestDto.title())
                 .content(requestDto.content())
-                .member(member)
+                .member(admin)
                 .build();
 
         noticeRepository.save(notice);
@@ -45,9 +47,17 @@ public class NoticeService {
      * 공지사항 수정
      */
     @Transactional
-    public NoticeUpdateResponseDto updateNotice(final Long noticeId, final NoticeUpdateRequestDto requestDto) {
+    public NoticeUpdateResponseDto updateNotice(
+            final String loginEmail,
+            final Long noticeId,
+            final NoticeUpdateRequestDto requestDto
+    ) {
+        final Member admin = getValidAdmin(loginEmail);
+
         final Notice notice = noticeRepository.findById(noticeId)
                 .orElseThrow(NoticeNotFoundException::new);
+
+        checkEqualAdmin(admin, notice);
 
         notice.update(requestDto.title(), requestDto.content());
 
@@ -82,10 +92,26 @@ public class NoticeService {
      * 공지사항 삭제
      */
     @Transactional
-    public void deleteNotice(final Long noticeId) {
+    public void deleteNotice(final String loginEmail, final Long noticeId) {
+        final Member admin = getValidAdmin(loginEmail);
+
         final Notice notice = noticeRepository.findById(noticeId)
                 .orElseThrow(NoticeNotFoundException::new);
 
+        checkEqualAdmin(admin, notice);
+
         noticeRepository.delete(notice);
+    }
+
+    // 관리자인지 확인
+    private Member getValidAdmin(final String loginEmail) {
+        return memberRepository.findByEmail(loginEmail)
+                .orElseThrow(() -> new UsernameNotFoundException("관리자를 찾을 수 없습니다."));
+    }
+
+    // 작성한 관리자인지 확인
+    private void checkEqualAdmin(final Member admin, final Notice notice) {
+        if (!notice.getMember().getId().equals(admin.getId()))
+            throw new NoticeInvalidException();
     }
 }
